+++ mux/src/tab.rs
@@ -714,8 +714,10 @@ impl Tab {
     }
 
     /// Swap the active pane with the specified pane_index
-    pub fn swap_active_with_index(&self, pane_index: usize) -> Option<()> {
-        self.inner.lock().swap_active_with_index(pane_index)
+    pub fn swap_active_with_index(&self, pane_index: usize, keep_focus: bool) -> Option<()> {
+        self.inner
+            .lock()
+            .swap_active_with_index(pane_index, keep_focus)
     }
 
     /// Computes the size of the pane that would result if the specified
@@ -1776,7 +1778,7 @@ impl TabInner {
         cell_dimensions(&self.size)
     }
 
-    fn swap_active_with_index(&mut self, pane_index: usize) -> Option<()> {
+    fn swap_active_with_index(&mut self, pane_index: usize, keep_focus: bool) -> Option<()> {
         let active_idx = self.get_active_idx();
         let mut pane = self.get_active_pane()?;
         log::trace!(
@@ -1822,7 +1824,11 @@ impl TabInner {
         }
 
         // And update focus
-        self.advise_focus_change(Some(pane));
+        if keep_focus {
+            self.set_active_idx(pane_index);
+        } else {
+            self.advise_focus_change(Some(pane));
+        }
         None
     }