+++ wezterm-gui/src/commands.rs
@@ -820,6 +820,17 @@ pub fn derive_command_from_key_assignment(action: &KeyAssignment) -> Option<Comm
             icon: Some("cod_multiple_windows"),
         },
         PaneSelect(PaneSelectArguments {
+            mode: PaneSelectMode::SwapWithActiveKeepFocus,
+            ..
+        }) => CommandDef {
+            brief: "Swap a pane with the active pane, keeping focus".into(),
+            doc: "Activates the pane selection UI".into(),
+            keys: vec![], // FIXME: find a new assignment
+            args: &[ArgType::ActivePane],
+            menubar: &["Window"],
+            icon: Some("cod_multiple_windows"),
+        },
+        PaneSelect(PaneSelectArguments {
             mode: PaneSelectMode::MoveToNewTab,
             ..
         }) => CommandDef {
@@ -2015,6 +2026,11 @@ fn compute_default_actions() -> Vec<KeyAssignment> {
         PaneSelect(PaneSelectArguments {
             alphabet: String::new(),
             mode: PaneSelectMode::SwapWithActive,
+            show_pane_ids: false,
+        }),
+        PaneSelect(PaneSelectArguments {
+            alphabet: String::new(),
+            mode: PaneSelectMode::SwapWithActiveKeepFocus,
             show_pane_ids: false,
         }),
         PaneSelect(PaneSelectArguments {